+++ server/domain/model/user.go
@@ -9,3 +9,13 @@ type User struct {
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
+
+// NewUser は、User を生成し、返す。
+func NewUser(id, name string) *User {
+	return &User{
+		ID:        id,
+		Name:      name,
+		CreatedAt: time.Now(),
+		UpdatedAt: time.Now(),
+	}
+}
